Rename AddTransaction class
